dom/canvas/test/webgl-mochitest/test_webgl_available.html

<!DOCTYPE HTML>
<html>
  <head>
    <meta charset='utf-8'/>
    <title>WebGL test: Check that WebGL works out-of-the-box.</title>
    <script src="/tests/SimpleTest/SimpleTest.js"></script>
    <link rel="stylesheet" href="/tests/SimpleTest/test.css">
  </head>
  <body>
    <script>

'use strict';
var c = document.createElement('canvas');
var gl = c.getContext('experimental-webgl');
ok(gl, 'Expected WebGL creation to succeed.');

    </script>
  </body>
</html>
